carga/requirements.txt
PyPDF2==3.0.1
Pygments==2.18.0
camelot-py==0.11.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
compact-json==1.8.1
contourpy==1.3.1
cramjam==2.9.1
cryptography==44.0.0
cycler==0.12.1
enum-tools==0.12.0
et_xmlfile==2.0.0
fonttools==4.55.3
ghostscript==0.7
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.3
mysql-connector-python==9.1.0
numbers-parser==4.14.1
numpy==2.2.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
pillow==11.0.0
protobuf==5.29.1
pycparser==2.22
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-snappy==0.7.3
pytz==2024.2
setuptools==75.6.0
sigfig==1.3.18
six==1.17.0
sortedcontainers==2.4.0
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
